+++ tests/parser/test_aircraft_beacon.py
@@ -20,13 +20,12 @@ class TestStringMethods(unittest.TestCase):
         self.assertEqual(aircraft_beacon['frequency_offset'], 51.2)
         self.assertEqual(aircraft_beacon['gps_status'], '4x5')
 
-    @unittest.skip("heared aircrafts not supported yet")
     def test_hear(self):
         aircraft_beacon = parse_ogn_aircraft_beacon("id0ADDA5BA -454fpm -1.1rot 8.8dB 0e +51.2kHz gps4x5 hear1084 hearB597 hearB598")
-        self.assertEqual(len(aircraft_beacon['heared_aircraft_addresses']), 3)
-        self.assertEqual(aircraft_beacon['heared_aircraft_addresses'][0], '1084')
-        self.assertEqual(aircraft_beacon['heared_aircraft_addresses'][1], 'B597')
-        self.assertEqual(aircraft_beacon['heared_aircraft_addresses'][2], 'B598')
+        self.assertEqual(len(aircraft_beacon['proximity']), 3)
+        self.assertEqual(aircraft_beacon['proximity'][0], '1084')
+        self.assertEqual(aircraft_beacon['proximity'][1], 'B597')
+        self.assertEqual(aircraft_beacon['proximity'][2], 'B598')
 
     def test_stealth(self):
         aircraft_beacon = parse_ogn_aircraft_beacon("id0ADD1234 -454fpm -1.1rot 8.8dB 0e +51.2kHz gps4x5 hear1084 hearB597 hearB598")